chore: maintenance commit! (05:07 PM - 21 November, 24)

+++ C++/tricky_sum_cf.cpp
@@ -12,15 +12,15 @@ int main() {
     long long sum;
 
     if (n % 2 == 0) {
-      sum = ((double)n / 2) * (n + 1);
+      sum = ((long double)n / 2) * (n + 1);
     } else {
-      sum = ((double)(n + 1) / 2) * (n);
+      sum = ((long double)(n + 1) / 2) * (n);
     }
 
     std::cout << sum << std::endl;
 
     for (int i = 0; i < n; i++) {
-      int num = pow(2, i);
+      long num = pow(2, i);
       if (num > n) {
         break;
       }
